BugFix [0000133]:
db allows the delete contact op to delete the main user chain COMPLETELY CORRUPTING the profile, now checks the raw offset isnt the user chain and denies the op.

This fix is related to the bugfix, stopping this stops addcontact using a weird hcontact value which when the user deletes corrupts the profile since it deletes the main user chain.


git-svn-id: http://miranda.googlecode.com/svn/trunk@285 83df23c2-51fd-11de-bb24-2d5e9b4af41d