Landing my current FT work. This is not finished, but good enough to be used.
Biggest changes are ability to download file batches (multiple files per single FT session) and the ability to send files (1 at a time for now) using the Yahoo 7 Ft protocol instead of the pre Yahoo 7 upload method. This one is something a lot of people were asking for. I am going to work on adding batch file sending but it's not critical at the moment and probaly not that common either.

git-svn-id: http://miranda.googlecode.com/svn/trunk@9136 83df23c2-51fd-11de-bb24-2d5e9b4af41d